+++ web/typescript/backend/AgendaEdit.ts
@@ -3,16 +3,14 @@
 declare let moment: any;
 
 export class AgendaEdit {
-    private hasChanged: boolean = false;
-    private $agendaform: JQuery;
     private readonly $agenda: JQuery;
     private readonly locale: string;
 
     private delAgendaItemStr = '<a href="#" class="delAgendaItem"><span class="glyphicon glyphicon-minus-sign"></span></a>';
 
     constructor(private $widget: JQuery) {
-        this.$agendaform = $('#agendaEditSavingHolder');
         this.$agenda = $('.motionListWithinAgenda');
+        this.locale = $("html").attr('lang');
 
         this.$agenda.addClass('agendaListEditing');
         this.$agenda.nestedSortable({
@@ -25,35 +23,17 @@ export class AgendaEdit {
             helper: 'clone',
             axis: 'y',
             update: () => {
-                this.showSaver();
-                $('ol.motionListWithinAgenda').trigger("antragsgruen:agenda-change");
+                this.$agenda.trigger("antragsgruen:agenda-change").trigger("antragsgruen:agenda-reordered");
             }
+        });
+        this.$agenda.on("antragsgruen:agenda-reordered", () => {
+            this.saveNewOrder();
         });
         this.$agenda.find('.agendaItem').each((i, el) => {
             this.prepareAgendaItem($(el));
         });
-        this.prepareAgendaList(this.$agenda, true);
-        this.$agenda.find('ol.agenda').each((i, el) => {
-            this.prepareAgendaList($(el), false);
-        });
 
-        this.locale = $("html").attr('lang');
-        this.$agenda.find(".input-group.time").datetimepicker({
-            locale: this.locale,
-            format: 'LT',
-            stepping: 5
-        });
-        this.$agenda.find(".input-group.date").each((i, el) => {
-            let preDate = null;
-            if ($(el).data("date")) {
-                preDate = moment($(el).data("date"), "YYYY-MM-DD", this.locale);
-            }
-            $(el).datetimepicker({
-                locale: this.locale,
-                format: 'dddd, Do MMMM YYYY',
-                defaultDate: preDate
-            });
-        });
+        this.prepareAgendaList(this.$agenda, true);
 
 
         this.$agenda.on('click', '.agendaItemAdder .addEntry', this.agendaItemAdd.bind(this));
@@ -62,38 +42,23 @@ export class AgendaEdit {
         this.$agenda.on('click', '.delAgendaItem', this.delAgendaItem.bind(this));
         this.$agenda.on('click', '.editAgendaItem', this.editAgendaItem.bind(this));
         this.$agenda.on('submit', '.agendaItemEditForm', this.submitSingleItemForm.bind(this));
-        this.$agenda.on('submit', '.agendaDateEditForm', this.submitDateItemForm.bind(this));
-        this.$agendaform.on("submit", this.submitCompleteForm.bind(this));
+        this.$agenda.on('submit', '.agendaDateEditForm', this.submitSingleItemForm.bind(this));
     }
 
     buildAgendaStruct($ol) {
         let items = [];
         $ol.children('.agendaItem').each((i, el) => {
-            const $li = $(el),
-                id = $li.attr('id').split('_');
+            const $li = $(el);
             if ($li.find('> div > .agendaDateEditForm').length > 0) {
-                let date = $li.find('> div > .agendaDateEditForm input[name=date]').val();
                 items.push({
                     'type': 'date',
-                    'id': id[1],
-                    'date': moment(date, 'dddd, Do MMMM YYYY', this.locale).format('L'),
-                    'title': $li.find('> div > .agendaDateEditForm input[name=title]').val(),
+                    'id': parseInt($li.data("id")),
                     'children': this.buildAgendaStruct($li.find('> ol'))
                 });
             } else {
-                const $form = $li.find('> div > .agendaItemEditForm');
-                let time = null;
-                if (this.$widget.find('.agendaItemAdder .showTimes input').prop("checked")) {
-                    time = $li.find('> div > .agendaItemEditForm input[name=time]').val();
-                }
                 items.push({
                     'type': 'std',
-                    'id': id[1],
-                    'time': time,
-                    'code': $form.find('input[name=code]').val(),
-                    'title': $form.find('input[name=title]').val(),
-                    'motionTypeId': $form.find('select[name=motionType]').val(),
-                    'inProposedProcedures': $form.find('.extraSettings input[name=inProposedProcedures]').prop('checked'),
+                    'id': parseInt($li.data("id")),
                     'children': this.buildAgendaStruct($li.find('> ol'))
                 });
             }
@@ -101,74 +66,90 @@ export class AgendaEdit {
         return items;
     }
 
-    submitDateItemForm(ev: Event) {
-        ev.preventDefault();
-        this.showSaver();
+    private saveNewOrder() {
+        const saveUrl = this.$widget.data("save-order") as string,
+            structure = this.buildAgendaStruct(this.$widget.find("> ol"));
 
-        let $li = $(ev.target).parents('li.agendaItem').first(),
-            $form = $(ev.target),
-            newTitle = $form.find('input[name=title]').val() as string,
-            newDate = $form.find('input[name=date]').val() as string,
-            fullString = '';
-        if (newDate) {
-            fullString += newDate;
-        }
-        if (newDate && newTitle) {
-            fullString += ': ';
-        }
-        if (newTitle) {
-            fullString += newTitle;
-        }
-        $li.removeClass('editing');
-        $li.find('> div > h3 .title').text(fullString);
-        $('ol.motionListWithinAgenda').trigger("antragsgruen:agenda-change");
+        $.post(saveUrl, {
+            '_csrf': $('input[name=_csrf]').val(),
+            'data': JSON.stringify(structure),
+        }, ret => {
+            if (!ret['success']) {
+                alert("Could not save: " + ret['message']);
+                return;
+            }
+            this.$agenda.trigger("antragsgruen:agenda-change");
+        });
     }
 
     submitSingleItemForm(ev: Event) {
         ev.preventDefault();
-        this.showSaver();
         let $li = $(ev.target).parents('li.agendaItem').first(),
+            $parentLi = $li.parents('li').first(),
+            position = $li.prevAll().length,
+            parentId = ($parentLi.length > 0 ? $parentLi.data('id') : null),
             $form = $(ev.target),
-            newTitle = $form.find('input[name=title]').val() as string,
-            newCode = $form.find('input[name=code]').val() as string,
-            newTime = $form.find('input[name=time]').val() as string;
-        $li.removeClass('editing');
-        $li.data('code', newCode);
-        $li.find('> div > h3 .code').text(newCode);
-        $li.find('> div > h3 .title').text(newTitle);
-        if (this.$widget.find('.agendaItemAdder .showTimes input').prop("checked") && newTime) {
-            if ($li.find('> div > h3 .time').length === 0) {
-                console.log("prepend");
-                $li.find('> div > h3').prepend('<span class="time"></span>');
-            }
-            $li.find('> div > h3 .time').text(newTime);
-        } else {
-            $li.find('> div > h3 .time').remove();
-        }
-        $('ol.motionListWithinAgenda').trigger("antragsgruen:agenda-change");
-    }
+            saveUrl = $li.data('save-url') as string;
 
-    submitCompleteForm() {
-        let data = this.buildAgendaStruct($('.motionListWithinAgenda'));
-        this.$agendaform.find('input[name=data]').val(JSON.stringify(data));
-        $(window).off("beforeunload", AgendaEdit.onLeavePage);
+        const saveData = {
+            'parentId': parentId,
+            'position': position,
+            'title': $form.find('input[name=title]').val() as string,
+        };
+
+        if ($li.hasClass('agendaItemDate')) {
+            saveData['type'] = 'date';
+            const date = ($form.find(".date") as any).datetimepicker("date");
+            if (date) {
+                saveData['date'] = date.format("YYYY-MM-DD");
+            } else {
+                saveData['date'] = null;
+            }
+        } else {
+            saveData['type'] = 'agendaItem';
+            saveData['inProposedProcedures'] = $form.find('input[name=inProposedProcedures]').prop('checked');
+            saveData['motionType'] = parseInt($form.find('select[name=motionType]').val() as string);
+            saveData['time'] = $form.find('input[name=time]').val() as string;
+            saveData['code'] = $form.find('input[name=code]').val() as string;
+        }
+
+        $.post(saveUrl, {
+            '_csrf': $('input[name=_csrf]').val(),
+            'data': JSON.stringify(saveData),
+        }, ret => {
+            if (!ret['success']) {
+                alert("Could not save: " + ret['message']);
+                return;
+            }
+            const $newItem = $(ret['html']);
+            $li.replaceWith($newItem);
+            this.prepareAgendaItem($newItem);
+            $('ol.motionListWithinAgenda').trigger("antragsgruen:agenda-change");
+        });
     }
 
     delAgendaItem(ev: Event) {
-        let $this = $(ev.target);
         ev.preventDefault();
+        let $li = $(ev.target).parents('li.agendaItem').first(),
+            delUrl = $li.data('del-url') as string;
         bootbox.confirm(__t("admin", "agendaDelEntryConfirm"), (result) => {
             if (result) {
-                this.showSaver();
-                $this.parents('li.agendaItem').first().remove();
-                $('ol.motionListWithinAgenda').trigger("antragsgruen:agenda-change");
+                $.post(delUrl, {
+                    '_csrf': $('input[name=_csrf]').val(),
+                }, ret => {
+                    if (!ret['success']) {
+                        alert("Could not delete: " + ret['message']);
+                        return;
+                    }
+                    $li.remove();
+                    $('ol.motionListWithinAgenda').trigger("antragsgruen:agenda-change");
+                });
             }
         });
     }
 
     editAgendaItem(ev: Event) {
         ev.preventDefault();
-        this.showSaver();
         let $li = $(ev.target).parents('li.agendaItem').first();
         $li.addClass('editing');
         $li.find('> div > .agendaItemEditForm input[name=code]').trigger("focus").trigger("select");
@@ -176,33 +157,21 @@ export class AgendaEdit {
 
     agendaItemAdd(ev: Event) {
         ev.preventDefault();
-        this.showSaver();
         let $newElement = $($('#agendaNewElementTemplate').val() as string),
             $adder = $(ev.target).parents('.agendaItemAdder').first();
         $adder.before($newElement);
         this.prepareAgendaItem($newElement);
-        this.prepareAgendaList($newElement.find('ol.agenda'), false);
         $newElement.find('.editAgendaItem').trigger('click');
         $newElement.find('.agendaItemEditForm input.code').trigger("focus");
-        $newElement.find(".input-group.time").datetimepicker({
-            locale: this.locale,
-            format: 'LT'
-        });
     }
 
     agendaDateAdd(ev: Event) {
         ev.preventDefault();
-        this.showSaver();
         let $newElement = $($('#agendaNewDateTemplate').val() as string),
             $adder = $(ev.target).parents('.agendaItemAdder').first();
         $adder.before($newElement);
         this.prepareAgendaItem($newElement);
-        this.prepareAgendaList($newElement.find('ol.agenda'), false);
         $newElement.find('.editAgendaItem').trigger('click');
-        $newElement.find(".input-group.date").datetimepicker({
-            locale: this.locale,
-            format: 'dddd, Do MMMM YYYY'
-        });
     }
 
     showTimesChanges() {
@@ -217,8 +186,29 @@ export class AgendaEdit {
         $item.find('> div').prepend('<span class="glyphicon glyphicon-resize-vertical moveHandle"></span>');
         $item.find('> div > h3').append('<a href="#" class="editAgendaItem"><span class="glyphicon glyphicon-pencil"></span></a>');
         $item.find('> div > h3').append(this.delAgendaItemStr);
-        $item.find('li.checkbox').on('click', (ev) => {
+        $item.find('> div li.checkbox').on('click', (ev) => {
             ev.stopPropagation();
+        });
+
+        $item.find('> ol.agenda').each((i, el) => {
+            this.prepareAgendaList($(el), false);
+        });
+
+        $item.find("> div .input-group.time").datetimepicker({
+            locale: this.locale,
+            format: 'LT'
+        });
+
+        $item.find("> div .input-group.date").each((i, el) => {
+            let preDate = null;
+            if ($(el).data("date")) {
+                preDate = moment($(el).data("date"), "YYYY-MM-DD", this.locale);
+            }
+            $(el).datetimepicker({
+                locale: this.locale,
+                format: 'dddd, Do MMMM YYYY',
+                defaultDate: preDate
+            });
         });
     }
 
@@ -236,17 +226,5 @@ export class AgendaEdit {
             $el.find(".showTimes input").prop("checked", true);
         }
         $list.append($el);
-    }
-
-    showSaver() {
-        $('#agendaEditSavingHolder').removeClass("hidden");
-        this.hasChanged = true;
-        if (!$("body").hasClass('testing')) {
-            $(window).on("beforeunload", AgendaEdit.onLeavePage);
-        }
-    }
-
-    public static onLeavePage(): string {
-        return __t("std", "leave_changed_page");
     }
 }